Revert "[pipelineX](feature) Support set operation operator (#25251)"
This reverts commit 7b22ae0c80b9a5d7e70c3d4e5148c97385638c74.
