+++ be/src/pipeline/exec/join_probe_operator.h
@@ -84,10 +84,8 @@ public:
 
     Status set_child(OperatorXPtr child) override {
         if (OperatorX<LocalStateType>::_child_x) {
-            // when there already (probe) child, others is build child.
             set_build_side_child(child);
         } else {
-            // first child which is probe side is in this pipeline
             OperatorX<LocalStateType>::_child_x = std::move(child);
         }
         return Status::OK();
