+++ be/src/pipeline/exec/set_probe_sink_operator.cpp
@@ -22,7 +22,6 @@
 #include <memory>
 
 #include "pipeline/exec/operator.h"
-#include "vec/common/hash_table/hash_table_set_probe.h"
 #include "vec/exec/vset_operation_node.h"
 
 namespace doris {
@@ -69,197 +68,5 @@ template class SetProbeSinkOperatorBuilder<true>;
 template class SetProbeSinkOperatorBuilder<false>;
 template class SetProbeSinkOperator<true>;
 template class SetProbeSinkOperator<false>;
-
-template <bool is_intersect>
-Status SetProbeSinkOperatorX<is_intersect>::sink(RuntimeState* state, vectorized::Block* in_block,
-                                                 SourceState source_state) {
-    RETURN_IF_CANCELLED(state);
-    CREATE_SINK_LOCAL_STATE_RETURN_IF_ERROR(local_state);
-    SCOPED_TIMER(local_state._probe_timer);
-
-    if (_cur_child_id > 1) {
-        CHECK(local_state._shared_state->probe_finished_children_index[_cur_child_id - 1])
-                << fmt::format("child with id: {} should be probed first", _cur_child_id);
-    }
-
-    auto probe_rows = in_block->rows();
-    if (probe_rows > 0) {
-        RETURN_IF_ERROR(_extract_probe_column(local_state, *in_block, local_state._probe_columns,
-                                              _cur_child_id));
-        RETURN_IF_ERROR(std::visit(
-                [&](auto&& arg) -> Status {
-                    using HashTableCtxType = std::decay_t<decltype(arg)>;
-                    if constexpr (!std::is_same_v<HashTableCtxType, std::monostate>) {
-                        vectorized::HashTableProbeX<HashTableCtxType, is_intersect>
-                                process_hashtable_ctx(local_state, probe_rows);
-                        return process_hashtable_ctx.mark_data_in_hashtable(local_state, arg);
-                    } else {
-                        LOG(FATAL) << "FATAL: uninited hash table";
-                    }
-                },
-                *local_state._shared_state->hash_table_variants));
-    }
-
-    if (source_state == SourceState::FINISHED) {
-        _finalize_probe(local_state);
-    }
-    return Status::OK();
-}
-
-template <bool is_intersect>
-Status SetProbeSinkLocalState<is_intersect>::init(RuntimeState* state, LocalSinkStateInfo& info) {
-    RETURN_IF_ERROR(PipelineXSinkLocalState<SetDependency>::init(state, info));
-    SCOPED_TIMER(profile()->total_time_counter());
-    SCOPED_TIMER(_open_timer);
-    _probe_timer = ADD_TIMER(_profile, "ProbeTime");
-
-    Parent& parent = _parent->cast<Parent>();
-
-    _child_exprs.resize(parent._child_exprs.size());
-    for (size_t i = 0; i < _child_exprs.size(); i++) {
-        RETURN_IF_ERROR(parent._child_exprs[i]->clone(state, _child_exprs[i]));
-    }
-
-    auto& child_exprs_lists = _shared_state->child_exprs_lists;
-    child_exprs_lists[parent._cur_child_id] = _child_exprs;
-
-    // Add the if check only for compatible with old optimiser
-    if (child_exprs_lists.size() > 1) {
-        _probe_columns.resize(child_exprs_lists[1].size());
-    }
-    return Status::OK();
-}
-
-template <bool is_intersect>
-Status SetProbeSinkOperatorX<is_intersect>::_extract_probe_column(
-        SetProbeSinkLocalState<is_intersect>& local_state, vectorized::Block& block,
-        vectorized::ColumnRawPtrs& raw_ptrs, int child_id) {
-    auto& build_not_ignore_null = local_state._shared_state->build_not_ignore_null;
-
-    for (size_t i = 0; i < _child_exprs.size(); ++i) {
-        int result_col_id = -1;
-        RETURN_IF_ERROR(_child_exprs[i]->execute(&block, &result_col_id));
-
-        block.get_by_position(result_col_id).column =
-                block.get_by_position(result_col_id).column->convert_to_full_column_if_const();
-        auto column = block.get_by_position(result_col_id).column.get();
-
-        if (auto* nullable = check_and_get_column<vectorized::ColumnNullable>(*column)) {
-            auto& col_nested = nullable->get_nested_column();
-            if (build_not_ignore_null[i]) { //same as build column
-                raw_ptrs[i] = nullable;
-            } else {
-                raw_ptrs[i] = &col_nested;
-            }
-
-        } else {
-            if (build_not_ignore_null[i]) {
-                auto column_ptr = make_nullable(block.get_by_position(result_col_id).column, false);
-                local_state._probe_column_inserted_id.emplace_back(block.columns());
-                block.insert(
-                        {column_ptr, make_nullable(block.get_by_position(result_col_id).type), ""});
-                column = column_ptr.get();
-            }
-
-            raw_ptrs[i] = column;
-        }
-    }
-    return Status::OK();
-}
-
-template <bool is_intersect>
-void SetProbeSinkOperatorX<is_intersect>::_finalize_probe(
-        SetProbeSinkLocalState<is_intersect>& local_state) {
-    auto& valid_element_in_hash_tbl = local_state._shared_state->valid_element_in_hash_tbl;
-    auto& hash_table_variants = local_state._shared_state->hash_table_variants;
-    auto& probe_finished_children_index = local_state._shared_state->probe_finished_children_index;
-
-    if (_cur_child_id != (local_state._shared_state->child_quantity - 1)) {
-        _refresh_hash_table(local_state);
-        if constexpr (is_intersect) {
-            valid_element_in_hash_tbl = 0;
-        } else {
-            std::visit(
-                    [&](auto&& arg) {
-                        using HashTableCtxType = std::decay_t<decltype(arg)>;
-                        if constexpr (!std::is_same_v<HashTableCtxType, std::monostate>) {
-                            valid_element_in_hash_tbl = arg.hash_table.size();
-                        }
-                    },
-                    *hash_table_variants);
-        }
-        local_state._probe_columns.resize(
-                local_state._shared_state->child_exprs_lists[_cur_child_id + 1].size());
-    } else {
-        local_state._dependency->set_ready_for_read();
-    }
-    probe_finished_children_index[_cur_child_id] = true;
-}
-
-template <bool is_intersect>
-void SetProbeSinkOperatorX<is_intersect>::_refresh_hash_table(
-        SetProbeSinkLocalState<is_intersect>& local_state) {
-    auto& valid_element_in_hash_tbl = local_state._shared_state->valid_element_in_hash_tbl;
-    auto& hash_table_variants = local_state._shared_state->hash_table_variants;
-    std::visit(
-            [&](auto&& arg) {
-                using HashTableCtxType = std::decay_t<decltype(arg)>;
-                if constexpr (!std::is_same_v<HashTableCtxType, std::monostate>) {
-                    if constexpr (std::is_same_v<typename HashTableCtxType::Mapped,
-                                                 vectorized::RowRefListWithFlags>) {
-                        HashTableCtxType tmp_hash_table;
-                        bool is_need_shrink =
-                                arg.hash_table.should_be_shrink(valid_element_in_hash_tbl);
-                        if (is_intersect || is_need_shrink) {
-                            tmp_hash_table.hash_table.init_buf_size(
-                                    valid_element_in_hash_tbl / arg.hash_table.get_factor() + 1);
-                        }
-
-                        arg.init_once();
-                        auto& iter = arg.iter;
-                        auto iter_end = arg.hash_table.end();
-                        std::visit(
-                                [&](auto is_need_shrink_const) {
-                                    while (iter != iter_end) {
-                                        auto& mapped = iter->get_second();
-                                        auto it = mapped.begin();
-
-                                        if constexpr (is_intersect) { //intersected
-                                            if (it->visited) {
-                                                it->visited = false;
-                                                tmp_hash_table.hash_table.insert(iter->get_value());
-                                            }
-                                            ++iter;
-                                        } else { //except
-                                            if constexpr (is_need_shrink_const) {
-                                                if (!it->visited) {
-                                                    tmp_hash_table.hash_table.insert(
-                                                            iter->get_value());
-                                                }
-                                            }
-                                            ++iter;
-                                        }
-                                    }
-                                },
-                                vectorized::make_bool_variant(is_need_shrink));
-
-                        arg.inited = false;
-                        if (is_intersect || is_need_shrink) {
-                            arg.hash_table = std::move(tmp_hash_table.hash_table);
-                        }
-                    } else {
-                        LOG(FATAL) << "FATAL: Invalid RowRefList";
-                    }
-                } else {
-                    LOG(FATAL) << "FATAL: uninited hash table";
-                }
-            },
-            *hash_table_variants);
-}
-
-template class SetProbeSinkLocalState<true>;
-template class SetProbeSinkLocalState<false>;
-template class SetProbeSinkOperatorX<true>;
-template class SetProbeSinkOperatorX<false>;
 
 } // namespace doris::pipeline
