+++ be/src/pipeline/exec/set_source_operator.cpp
@@ -48,125 +48,4 @@ template class SetSourceOperatorBuilder<false>;
 template class SetSourceOperator<true>;
 template class SetSourceOperator<false>;
 
-template <bool is_intersect>
-Status SetSourceLocalState<is_intersect>::init(RuntimeState* state, LocalStateInfo& info) {
-    RETURN_IF_ERROR(PipelineXLocalState<SetDependency>::init(state, info));
-    _pull_timer = ADD_TIMER(_runtime_profile, "PullTime");
-    return Status::OK();
-}
-
-template <bool is_intersect>
-Status SetSourceOperatorX<is_intersect>::get_block(RuntimeState* state, vectorized::Block* block,
-                                                   SourceState& source_state) {
-    RETURN_IF_CANCELLED(state);
-    CREATE_LOCAL_STATE_RETURN_IF_ERROR(local_state);
-    SCOPED_TIMER(local_state._pull_timer);
-    _create_mutable_cols(local_state, block);
-    auto st = std::visit(
-            [&](auto&& arg) -> Status {
-                using HashTableCtxType = std::decay_t<decltype(arg)>;
-                if constexpr (!std::is_same_v<HashTableCtxType, std::monostate>) {
-                    return _get_data_in_hashtable<HashTableCtxType>(
-                            local_state, arg, block, state->batch_size(), source_state);
-                } else {
-                    LOG(FATAL) << "FATAL: uninited hash table";
-                }
-            },
-            *local_state._shared_state->hash_table_variants);
-    RETURN_IF_ERROR(st);
-    RETURN_IF_ERROR(vectorized::VExprContext::filter_block(local_state._conjuncts, block,
-                                                           block->columns()));
-    local_state.reached_limit(block, source_state);
-    return Status::OK();
-}
-
-template <bool is_intersect>
-void SetSourceOperatorX<is_intersect>::_create_mutable_cols(
-        SetSourceLocalState<is_intersect>& local_state, vectorized::Block* output_block) {
-    local_state._mutable_cols.resize(local_state._shared_state->left_table_data_types.size());
-    bool mem_reuse = output_block->mem_reuse();
-
-    for (int i = 0; i < local_state._shared_state->left_table_data_types.size(); ++i) {
-        if (mem_reuse) {
-            local_state._mutable_cols[i] =
-                    std::move(*output_block->get_by_position(i).column).mutate();
-        } else {
-            local_state._mutable_cols[i] =
-                    (local_state._shared_state->left_table_data_types[i]->create_column());
-        }
-    }
-}
-
-template <bool is_intersect>
-template <typename HashTableContext>
-Status SetSourceOperatorX<is_intersect>::_get_data_in_hashtable(
-        SetSourceLocalState<is_intersect>& local_state, HashTableContext& hash_table_ctx,
-        vectorized::Block* output_block, const int batch_size, SourceState& source_state) {
-    hash_table_ctx.init_once();
-    int left_col_len = local_state._shared_state->left_table_data_types.size();
-    auto& iter = hash_table_ctx.iter;
-    auto block_size = 0;
-
-    if constexpr (std::is_same_v<typename HashTableContext::Mapped,
-                                 vectorized::RowRefListWithFlags>) {
-        for (; iter != hash_table_ctx.hash_table.end() && block_size < batch_size; ++iter) {
-            auto& value = iter->get_second();
-            auto it = value.begin();
-            if constexpr (is_intersect) {
-                if (it->visited) { //intersected: have done probe, so visited values it's the result
-                    _add_result_columns(local_state, value, block_size);
-                }
-            } else {
-                if (!it->visited) { //except: haven't visited values it's the needed result
-                    _add_result_columns(local_state, value, block_size);
-                }
-            }
-        }
-    } else {
-        return Status::InternalError("Invalid RowRefListType!");
-    }
-
-    if (iter == hash_table_ctx.hash_table.end()) {
-        source_state = SourceState::FINISHED;
-    }
-    if (!output_block->mem_reuse()) {
-        for (int i = 0; i < left_col_len; ++i) {
-            output_block->insert(vectorized::ColumnWithTypeAndName(
-                    std::move(local_state._mutable_cols[i]),
-                    local_state._shared_state->left_table_data_types[i], ""));
-        }
-    } else {
-        local_state._mutable_cols.clear();
-    }
-
-    return Status::OK();
-}
-
-template <bool is_intersect>
-void SetSourceOperatorX<is_intersect>::_add_result_columns(
-        SetSourceLocalState<is_intersect>& local_state, vectorized::RowRefListWithFlags& value,
-        int& block_size) {
-    auto& build_col_idx = local_state._shared_state->build_col_idx;
-    auto& build_blocks = local_state._shared_state->build_blocks;
-
-    auto it = value.begin();
-    for (auto idx = build_col_idx.begin(); idx != build_col_idx.end(); ++idx) {
-        auto& column = *build_blocks[it->block_offset].get_by_position(idx->first).column;
-        if (local_state._mutable_cols[idx->second]->is_nullable() xor column.is_nullable()) {
-            DCHECK(local_state._mutable_cols[idx->second]->is_nullable());
-            ((vectorized::ColumnNullable*)(local_state._mutable_cols[idx->second].get()))
-                    ->insert_from_not_nullable(column, it->row_num);
-
-        } else {
-            local_state._mutable_cols[idx->second]->insert_from(column, it->row_num);
-        }
-    }
-    block_size++;
-}
-
-template class SetSourceLocalState<true>;
-template class SetSourceLocalState<false>;
-template class SetSourceOperatorX<true>;
-template class SetSourceOperatorX<false>;
-
 } // namespace doris::pipeline
